On a fresh os, the source compilation is failing due to missing packages like flex bison for test like iperf, uperf and netperf

configure: error: Neither flex nor lex was found
configure: error: ./configure failed for libpcap

Signed-off-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>