@abdhaleegit 2023-01-18
before fix
````
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/uperf_test.py:Uperf.test;run-mtu-1500-224a: STARTED
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/uperf_test.py:Uperf.test;run-mtu-1500-224a:  ERROR: 'Uperf' object has no attribute 'obj' (31.42 s)
```
```
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/iperf_test.py:Iperf.test;run-mtu-1500-00ae: STARTED
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/iperf_test.py:Iperf.test;run-mtu-1500-00ae:  ERROR: 'Iperf' object has no attribute 'obj' (11.42 s)
 ```
 
 After fix 
 ```
  (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/uperf_test.py:Uperf.test;run-mtu-1500-f4bd: STARTED
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/uperf_test.py:Uperf.test;run-mtu-1500-f4bd:  FAIL: FAIL: Throughput Actual - 8.74%, Expected - 50%, Throughput Actual value - 874Mb/sec  (50.39 s)
 (2/9) /home/OpTest/avocado-f
 ```
 ```
  (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/iperf_test.py:Iperf.test;run-mtu-1500-c26a: STARTED
 (1/9) /home/OpTest/avocado-fvt-wrapper/tests/avocado-misc-tests/io/net/iperf_test.py:Iperf.test;run-mtu-1500-c26a:  FAIL: FAIL: Throughput Actual - 9.44%, Expected - 50%, Throughput Actual value - 944Mb/sec  (33.47 s)
 ```